+++ src/testCars.java
@@ -1,8 +1,4 @@
-import problem2.Sedan;
-import problem2.Sports;
-import problem2.Cars;
-import java.util.ArrayList;
-
+import java.util.*;
 
 public class testCars {
